test(electron): split sqlite ipc contract fixtures
