+++ apps/electron-backend/tsconfig.app.json
@@ -7,9 +7,11 @@
     "files": ["../../global.d.ts"],
     "exclude": [
         "**/*.spec.ts",
+        "**/*.spec-data.ts",
         "**/*.test.ts",
         "jest.config.ts",
         "src/**/*.spec.ts",
+        "src/**/*.spec-data.ts",
         "src/**/*.test.ts"
     ],
     "include": ["**/*.ts"]
